refactor: Rename images.rs to migrations.rs for better organization
- Rename src-tauri/src/images.rs to src-tauri/src/migrations.rs
- Update module imports in lib.rs to reflect new naming
- Improve code organization by having migrations in appropriately named file
